# contract/openapi.yaml
openapi: 3.1.0
info:
  title: SaaS Starter API
  version: 1.0.0
  description: |
    Shared API contract for the SaaS Starter Kit.
    All backend implementations (Go, Python, Node.js, etc.) must conform to this spec.
servers:
  - url: http://localhost:8080
    description: Default development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth ──────────────────────────────────────────────
...
    AppSetting:
      type: object
      required: [id, key, value]
      properties:
        id:
          type: integer
        key:
          type: string
        value:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /api/admin/settings:
    get:
      tags: [Admin - Settings]
      summary: List all app settings
      operationId: getAllSettings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppSetting"
  /api/admin/settings/{key}:
    get:
      tags: [Admin - Settings]
      summary: Get a specific setting
      operationId: getSetting
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Setting value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSetting"
    put:
      tags: [Admin - Settings]
      summary: Update a specific setting
      operationId: updateSetting
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: string
      responses:
        "200":
          description: Setting updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSetting"
  /api/admin/settings/email-verification:
    get:
      tags: [Admin - Settings]
      summary: Get email verification setting
      operationId: getEmailVerificationSetting
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Email verification status
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled:
                    type: boolean
    put:
      tags: [Admin - Settings]
      summary: Update email verification setting
      operationId: updateEmailVerificationSetting
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        "200":
          description: Setting updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/admin/settings/2fa:
    get:
      tags: [Admin - Settings]
      summary: Get 2FA enforcement setting
      operationId: get2FASetting
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 2FA setting
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled:
                    type: boolean
    put:
      tags: [Admin - Settings]
      summary: Update 2FA enforcement setting
      operationId: update2FASetting
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        "200":
          description: Setting updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"